requirements.txt
boto3
botocore
gradio_client
Pillow
requests
tqdm
python-dotenv
distinctipy
opencv-python
